Copy _v2 version over original; we don't need two versions.
